extension methods for Editor
