+++ Assets/Project/Scripts/Editor/EditorExtension.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6f703c56182b444fc96737cb2545a469
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
